requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.3.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.1
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.3
fitz==0.0.1.dev2
Flask==3.0.2
frontend==0.0.3
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.6
image==1.5.33
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
looseversion==1.3.0
lxml==5.2.0
MarkupPy==1.14
MarkupSafe==2.1.5
multidict==6.0.5
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==24.0
pandas==2.2.1
pathlib==1.0.1
pdf2image==1.17.0
pillow==10.3.0
prov==2.0.0
pycparser==2.22
pydot==2.0.0
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
pypdf==4.2.0
pyperclip==1.8.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.0
pytils==0.4.1
pytz==2023.3.post1
pyxnat==1.6.2
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.2
rdflib==7.0.0
reportlab==4.0.9
requests==2.31.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.37.2
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
tools==0.1.9
traits==6.3.2
twilio==9.0.5
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
uvicorn==0.29.0
webencodings==0.5.1
Werkzeug==3.0.2
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
